v1.1
1.替换了新的RF驱动
2.删除了多余的打印输出
3.

+++ applications/FlashWork.h
@@ -53,7 +53,7 @@ uint32_t Flash_Get_Moto_Flag(void);
 uint32_t Flash_Get_Door_Nums(void);
 uint32_t Flash_Get_Gateway_Nums(void);
 uint8_t Update_Device_Bat(uint32_t Device_ID,uint8_t bat);
-uint8_t Update_Device_Rssi(uint32_t Device_ID,uint8_t rssi);//更新Rssi;
+uint8_t Update_Device_Rssi(uint32_t Device_ID,int rssi);//更新Rssi;
 void Offline_React(uint32_t ID);
 uint8_t Device_RssiGet(uint32_t Device_ID);
 void Device_RssiChange(uint32_t Device_ID,uint8_t value);
